fixed `IsActorBaseSubclass`
